% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossover.R
\name{crossover}
\alias{crossover}
\title{crossover()}
\usage{
crossover(genes, parents)
}
\arguments{
\item{genes}{n by P matrix; P columns of genes with length n (composed of 1s and 0s)}

\item{parents}{2 by P marix; P columns of indexes}
}
\description{
This function receives input from the output
of selection function, and do crossovers on the pairs of parents
}
\author{
Fernando Machado, Jiayang Nie
}
