% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossover.R
\name{crossover_helper}
\alias{crossover_helper}
\title{crossover_helper()}
\usage{
crossover_helper(parent.ind, genes)
}
\arguments{
\item{parent.ind}{list of 2 integers; indexes of the parents}

\item{genes}{n by P matrix; P columns of genes with length n (composed of 1s and 0s)}
}
\description{
A helper function for crossover
}
\author{
Fernando Machado, Jiayang Nie
}
