% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{mutation}
\alias{mutation}
\title{mutation()}
\usage{
mutation(genes, mutation_rate = NULL)
}
\arguments{
\item{genes}{n by P matrix; P columns of genes with length n (composed of 1s and 0s)}

\item{mutation_rate}{double; indicate the mutation rate}
}
\description{
This function receives input from the output
of crossover function, and do mutation on offsprings. By default,
Mutation_rate is set to 1/C where C is the gene length.
}
\author{
Fernando Machado, Jiayang Nie
}
