% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{random_mutate}
\alias{random_mutate}
\title{random_mutate()}
\usage{
random_mutate(gene)
}
\arguments{
\item{gene}{list of integers; a single gene of genes}
}
\description{
Randomly mutate a gene
}
\author{
Fernando Machado, Jiayang Nie
}
