% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection.R
\name{selection}
\alias{selection}
\title{selection()}
\usage{
selection(genes, fit_values, type = "tournament")
}
\arguments{
\item{genes}{n by P matrix; P columns of genes with length n (composed of 1s and 0s)}

\item{fit_values}{list of doubles; list of fitness values, with the order of @param genes}

\item{type}{string; type could be "tournament" or "rank". tournament is selection by tuornament method, and rank is selection by rank}
}
\description{
This function receives input from the output
of evaluation function, and select the best P pairs of genes as parents,
stored as column indexes for the matrix. The output will be a 2 by P matrix,
with each column as the parents' index in the original matrix
}
\author{
Jiayang Nie
}
