% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection.R
\name{tournament_vectorization}
\alias{tournament_vectorization}
\title{tournament_vectorization()}
\usage{
tournament_vectorization(x, fit_values)
}
\arguments{
\item{x}{list of 6 indexes; each index can be mapped to a fitness value, and a gene}

\item{fit_values}{list of doubles; list of fitness values passed in, with the order of @param genes}
}
\description{
This function is a helper function for the apply vectorization in selection function
}
\author{
Jiayang Nie
}
